* Improved extract of cable type. * Some minor check are added when generating report documents. * Added formulas to excel report.

+++ Source/CableManager.Services/Helpers/CalculationHelper.cs
@@ -13,9 +13,9 @@ namespace CableManager.Services.Helpers
          return price * rebateQuota * quantity;
       }
 
-      public static float CalculatePriceWithVat(float price, int valueAddedTax)
+      public static float CalculatePriceWithVat(float price, float valueAddedTax)
       {
-         float vatQuota = ((float)(100 + valueAddedTax) / 100);
+         float vatQuota = ((100 + valueAddedTax) / 100);
 
          return price * vatQuota;
       }
